# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Harém Brasil API
  version: 1.0.0
  description: API backend para a plataforma Harém Brasil

servers:
  - url: http://localhost:40080
    description: Desenvolvimento local

paths:
  /api/v1/auth/register:
    post:
      operationId: register
      summary: Registar novo usuário
      tags: [Autenticação]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            examples:
              valid:
                summary: Registo válido
                value:
                  email: user@example.com
                  screen_name: cooluser
                  password: SecurePass123!
                  accept_terms_version: "1.0"
      responses:
        "201":
          description: Utilizador criado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: JSON inválido ou corpo malformado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Email ou screen_name já existe
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "422":
          description: Erro de validação (campos obrigatórios ou formato inválido)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "429":
          description: Rate limit excedido
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

  /api/v1/auth/login:
    post:
      operationId: login
      summary: Autenticar usuário com email e senha
      tags: [Autenticação]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              valid:
                summary: Credenciais válidas
                value:
                  email: user@example.com
                  password: SecurePass123!
      responses:
        "200":
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: JSON inválido ou corpo malformado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Credenciais inválidas
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "422":
          description: Erro de validação (campos obrigatórios ou formato inválido)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "429":
          description: Rate limit excedido
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

  /api/v1/creator/catalog:
    get:
      operationId: listCreatorCatalog
      tags: [Criador]
      security:
        - bearerAuth: []
      summary: Listar itens do catálogo do criador autenticado
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Timestamp RFC3339 copiado de next_cursor da página anterior (paginação por created_at)
      responses:
        "200":
          description: Lista paginada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CursorPageCreatorCatalog"
              examples:
                empty:
                  summary: Sem itens
                  value:
                    data: []
                    has_more: false
        "400":
          description: Cursor inválido
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Não autenticado
        "403":
          description: Papel insuficiente (requer creator ou admin)

    post:
      operationId: createCreatorCatalogItem
      tags: [Criador]
      security:
        - bearerAuth: []
      summary: Criar item pago no catálogo (ligado ao perfil do criador)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatorCatalogCreateRequest"
            examples:
              pack:
                summary: Item com mídia
                value:
                  title: Pack fotos exclusivas
                  description: Coleção em alta resolução.
                  price_cents: 2990
                  currency: BRL
                  visibility: subscribers
                  media:
                    - https://cdn.example.com/preview1.jpg
      responses:
        "201":
          description: Item criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorCatalogItem"
        "400":
          description: Validação ou JSON inválido
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Não autenticado
        "403":
          description: Papel insuficiente

  /api/v1/creator/catalog/{item_id}:
    parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: patchCreatorCatalogItem
      tags: [Criador]
      security:
        - bearerAuth: []
      summary: Atualizar item (whitelist; só dono ou admin)
      description: |
        Campos omitidos não são alterados. Outro criador recebe 404 (BOLA).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatorCatalogPatchRequest"
            examples:
              price:
                summary: Só preço
                value:
                  price_cents: 3490
      responses:
        "200":
          description: Item atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorCatalogItem"
        "400":
          description: Nenhum campo ou validação falhou
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Não autenticado
        "403":
          description: Papel insuficiente
        "404":
          description: Item inexistente ou não autorizado para este usuário

    delete:
      operationId: deleteCreatorCatalogItem
      tags: [Criador]
      security:
        - bearerAuth: []
      summary: Remover item do catálogo (DELETE físico na base de dados)
      responses:
        "204":
          description: Linha removida de creator_catalog
        "401":
          description: Não autenticado
        "403":
          description: Papel insuficiente
        "404":
          description: Item inexistente ou não autorizado
        "409":
          description: Item referenciado por pedidos (creator_orders); não pode apagar
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

components:
  schemas:
    RegisterRequest:
      type: object
      required: [email, screen_name, password, accept_terms_version]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
          description: Email do usuário
          examples:
            - user@example.com
        screen_name:
          type: string
          minLength: 2
          maxLength: 64
          description: Nome de exibição do usuário
          examples:
            - cooluser
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[\\p{P}\\p{S}]).{8,}$"
          description: Senha do usuário (mínimo 8 caracteres, 1 maiúscula, 1 minúscula, 1 dígito, 1 especial)
        accept_terms_version:
          type: string
          maxLength: 32
          description: Versão dos termos aceites pelo usuário
          examples:
            - "1.0"
      additionalProperties: false

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
          description: Email do usuário
          examples:
            - user@example.com
        password:
          type: string
          format: password
          maxLength: 128
          description: Senha do usuário
      additionalProperties: false

    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT de acesso (expira em 15 minutos)
        access_expires_in:
          type: integer
          description: Duração do access token em segundos
          examples:
            - 900
        refresh_token:
          type: string
          description: Token de renovação (expira em 7 dias)
        refresh_expires_in:
          type: integer
          description: Duração do refresh token em segundos
          examples:
            - 604800
        token_type:
          type: string
          enum: [Bearer]
        expires_at:
          type: string
          format: date-time
          description: Timestamp de expiração do access token
        user:
          $ref: "#/components/schemas/UserPublic"
      required: [access_token, access_expires_in, refresh_token, refresh_expires_in, token_type, expires_at, user]

    UserPublic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        screen_name:
          type: string
          description: Nome de exibição do usuário
        email:
          type: string
          format: email
        role:
          type: string
          enum: [guest, user, creator, moderator, admin]
        avatar_url:
          type: string
          format: uri
        bio:
          type: string
        created_at:
          type: string
          format: date-time
      required: [id, screen_name, role, created_at]

    ProblemDetail:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
      required: [title, status]

    ValidationError:
      allOf:
        - $ref: "#/components/schemas/ProblemDetail"
        - type: object
          properties:
            extensions:
              type: object
              properties:
                fields:
                  type: object
                  additionalProperties:
                    type: string

    CursorPageCreatorCatalog:
      type: object
      additionalProperties: false
      required: [data, has_more]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CreatorCatalogItem"
        next_cursor:
          type: string
        has_more:
          type: boolean

    CreatorCatalogItem:
      type: object
      additionalProperties: false
      required:
        [id, title, description, price_cents, currency, visibility, media, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 10000
        price_cents:
          type: integer
          minimum: 0
          maximum: 999999999
          description: Preço em cêntimos (inteiro)
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
          description: ISO 4217
          examples: [BRL, USD]
        visibility:
          type: string
          enum: [public, subscribers, premium]
        media:
          type: array
          maxItems: 20
          items:
            type: string
            maxLength: 2048
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreatorCatalogCreateRequest:
      type: object
      additionalProperties: false
      required: [title, price_cents, currency, visibility]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 10000
        price_cents:
          type: integer
          minimum: 0
          maximum: 999999999
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO 4217 (normalizado para maiúsculas na API)
        visibility:
          type: string
          enum: [public, subscribers, premium]
        media:
          type: array
          maxItems: 20
          items:
            type: string
            maxLength: 2048

    CreatorCatalogPatchRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 10000
        price_cents:
          type: integer
          minimum: 0
          maximum: 999999999
        currency:
          type: string
          minLength: 3
          maxLength: 3
        visibility:
          type: string
          enum: [public, subscribers, premium]
        media:
          type: array
          maxItems: 20
          items:
            type: string
            maxLength: 2048

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT HS256 (issuer harem-api); rotas de criador exigem role creator ou admin.